adicionando pydantic
