+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "api-workshop02"
+version = "0.1.0"
+description = ""
+authors = ["Luciano Filho <lvgalvaofilho@gmail.com>"]
+readme = "README.md"
+packages = [{ include = "api_workshop02" }]
+
+[tool.poetry.dependencies]
+python = "^3.11"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.taskipy.tasks]
+run = "uvicorn app.main:app --reload"
+test = "pytest -v"
+kill = "kill -9 $(lsof -t -i:8000)"
